--- Starlet_BE/src/main/java/com/example/starlet_be/domains/user/reqdto/UserReqDto.java
@@ -1,30 +1,0 @@
-package com.example.starlet_be.domains.user.reqdto;
-
-import com.example.starlet_be.domains.user.entity.User;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
-
-@Data
-public class UserReqDto {
-    @NotBlank
-    private String nickname;
-
-    @NotBlank
-    private String password;
-
-    @NotBlank
-    @Email
-    private String email;
-
-    public User toEntity(String encodedPassword) {
-        return User.builder()
-                .nickname(nickname)
-                .password(encodedPassword)
-                .email(email)
-                .build();
-        // 초기에는 이메일 인증이 false 인 상태로 진행됨
-        // 위에 인텔리제이 자동완성인데 저건 뭔지 봐야할 듯 : 빌더의 의도에 맞게 된 문법이라 저렇게 쓰는게 맞음
-//        return new User(username, password, email);
-    }
-}
